+++ sync_types/src/udf_path.rs
@@ -3,6 +3,9 @@ use std::{
     str::FromStr,
 };
 
+#[cfg(any(test, feature = "testing"))]
+use proptest::prelude::*;
+
 use super::module_path::{
     CanonicalizedModulePath,
     ModulePath,